+++ drop-storage/src/error.rs
@@ -3,5 +3,7 @@ pub enum Error {
     #[error("Storage error: {0}")]
     InternalError(String),
     #[error("DB Error: {0}")]
-    DBError(#[from] sqlx::Error),
+    DBError(#[from] rusqlite::Error),
+    #[error("DB Connection Error: {0}")]
+    DBConnectionError(#[from] r2d2::Error),
 }